ArticlesDrawer refactored
