+++ core/getRootArticles.js
@@ -1,0 +1,17 @@
+import gql from "graphql-tag";
+
+const GET_ROOT_ARTICLES = gql`
+  query getRootArticles {
+    getRootArticles {
+      id
+      title
+      icon
+      content
+      parent {
+        id
+      }
+    }
+  }
+`;
+
+export default GET_ROOT_ARTICLES;
